image_search/requirements.txt
colorgram.py==1.2.0
Django==3.2.4
eventlet==0.31.0
fiftyone==0.9.3
Pillow==8.2.0
tqdm==4.61.0
translate==3.5.0

